# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IUT Cafeteria API
  description: |
    API for the IUT Cafeteria ordering system.

    ## Authentication
    All endpoints (except /login, /register, /health) require JWT authentication.
    Pass the token via:
    - `Authorization: Bearer <token>` header
    - `token` httpOnly cookie

    ## Correlation IDs
    All requests are assigned an `x-correlation-id` header for distributed tracing.
  version: 1.0.0
  contact:
    name: IUT Cafeteria Team
    email: admin@iut-cafeteria.app
servers:
  - url: http://localhost:3000
    description: Order Gateway (Development)
  - url: http://localhost:3001
    description: Identity Provider (Development)
  - url: http://localhost:3002
    description: Stock Service (Development)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: token
  schemas:
...
    Order:
      type: object
      properties:
        order_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, STOCK_VERIFIED, IN_KITCHEN, READY, PICKED_UP, FAILED, CANCELLED]
        amount:
          type: number
          format: float
        item_name:
          type: string
        created_at:
          type: string
          format: date-time
...
  parameters:
...
    IdempotencyKey:
      name: x-idempotency-key
      in: header
      description: Idempotency key for safe retries
      schema:
        type: string
        format: uuid
paths:
...
  /menu/{itemId}/reviews:
    get:
      tags: [Menu]
      summary: Get reviews for a menu item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Reviews list

  # ─── Orders ───────────────────────────────────────────────────
  /orders:
    get:
      tags: [Orders]
      summary: Get order history
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Order history with pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer

    post:
      tags: [Orders]
      summary: Place a new order
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itemId]
              properties:
                itemId:
                  type: integer
                type:
                  type: string
                  enum: [dinner, iftar]
                  default: dinner
                scheduledPickupTime:
                  type: string
                  format: date-time
      responses:
        '202':
          description: Order accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  orderId:
                    type: string
                    format: uuid
                  message:
                    type: string
                  item:
                    type: string
        '402':
          description: Insufficient balance
        '409':
          description: Out of stock
  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get order details
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
  /orders/verify-qr:
    post:
      tags: [Orders]
      summary: Verify QR code for pickup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [qrCode]
              properties:
                qrCode:
                  type: string
      responses:
        '200':
          description: Order verified
        '404':
          description: Invalid QR code
        '409':
          description: Already picked up

  # ─── Wallet ───────────────────────────────────────────────────
...
  /reviews:
    post:
      tags: [Orders]
      summary: Submit order review
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderId, rating]
              properties:
                orderId:
                  type: string
                  format: uuid
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
      responses:
        '200':
          description: Review submitted

  # ─── Admin ────────────────────────────────────────────────────
...